requirements0.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
django-model-utils==5.0.0
django-report-tools==0.2.2
django-slick-reporting==1.3.1
djangorestframework==3.15.2
mysqlclient==2.2.5
pathlib==1.0.1
pillow==11.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
Unipath==1.1
